added basic gui for settings

+++ core/src/io/piotrjastrzebski/playground/tiledgentest/MapData.java
@@ -10,6 +10,6 @@ public class MapData {
 	public float water;
 
 	public int largestFeature;
-	public double persistence;
+	public float persistence;
 	public long seed;
 }
